refactor(core): move wait_random and rdi into separate file
