+++ core/SConscript.firmware
@@ -333,6 +333,7 @@ SOURCE_TREZORHAL = [
     'embed/trezorhal/flash.c',
     'embed/trezorhal/mini_printf.c',
     'embed/trezorhal/mpu.c',
+    'embed/trezorhal/random_delays.c',
     'embed/trezorhal/rng.c',
     'embed/trezorhal/sbu.c',
     'embed/trezorhal/sdcard.c',
@@ -349,9 +350,6 @@ SOURCE_TREZORHAL = [
 ]
 
 if FEATURE_FLAGS["RDI"]:
-    SOURCE_TREZORHAL += [
-        'embed/trezorhal/rdi.c',
-    ]
     CPPDEFINES_MOD += ['RDI']
 
 if FEATURE_FLAGS["SYSTEM_VIEW"]:
