+++ core/embed/unix/random_delays.h
@@ -17,21 +17,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef __TREZORHAL_COMMON_H__
-#define __TREZORHAL_COMMON_H__
+#ifndef __TREZORHAL_RANDOM_DELAYS_H__
+#define __TREZORHAL_RANDOM_DELAYS_H__
 
-#include "secbool.h"
-
-void __fatal_error(const char *expr, const char *msg, const char *file,
-                   int line, const char *func);
-void error_shutdown(const char *line1, const char *line2, const char *line3,
-                    const char *line4);
-
-#define ensure(expr, msg) \
-  (((expr) == sectrue)    \
-       ? (void)0          \
-       : __fatal_error(#expr, msg, __FILE__, __LINE__, __func__))
-
-#define hal_delay(ms) (void)ms;
-
+void wait_random(void);
 #endif
