+++ storage/tests/c/common.c
@@ -23,8 +23,6 @@
 
 #include "common.h"
 
-void wait_random(void) {}
-
 void __shutdown(void) {
   printf("SHUTDOWN\n");
   exit(3);
