+++ storage/tests/c/random_delays.c
@@ -17,21 +17,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef __TREZORHAL_COMMON_H__
-#define __TREZORHAL_COMMON_H__
+#include "random_delays.h"
 
-#include "secbool.h"
-
-void __fatal_error(const char *expr, const char *msg, const char *file,
-                   int line, const char *func);
-void error_shutdown(const char *line1, const char *line2, const char *line3,
-                    const char *line4);
-
-#define ensure(expr, msg) \
-  (((expr) == sectrue)    \
-       ? (void)0          \
-       : __fatal_error(#expr, msg, __FILE__, __LINE__, __func__))
-
-#define hal_delay(ms) (void)ms;
-
-#endif
+void wait_random(void) {}
